ellipses kinda work
We have the eigenvectors of the covariance matrix,
which are the (directions of) the major and minor axis,
and we have the rotation angle.

The major/minor are not scaled properly; for now,
we naively take 0.7 * range (for either F1 or F2),
but the axes are in neither F1 or F2 direction.

The SVG output seems OK, as the plot axes have the same scale.
In PS though, the ellipses are even wrongly tilted, as the
Hz1/Hz2 scaling further screws the major/minor of the ellipse.